feat(daemon-ui): brand mark + theme/accent controls on the command center
The daemon surfaces had none of the PR-review chrome — no DadMark, no theme toggle.
Combined with per-origin localStorage (the daemon's :4319 can't see prefs saved on
the PR server's random port) and an 'auto' default, a dark OS stranded the command
center in dark with no way out. Extract ThemeToggle + AccentPicker out of AppBar into
shared, store-driven components and drop them — plus the brand mark — onto the command
center and the per-unit drill-in. AppBar now consumes the same two controls.

Adds lib/theme.ts (single source for the Theme type + cycleTheme, TDD'd).
