+++ packages/web/src/lib/__tests__/theme.test.ts
@@ -1,0 +1,10 @@
+import { describe, expect, it } from 'vitest';
+import { cycleTheme } from '../theme';
+
+describe('cycleTheme', () => {
+  it('cycles light → dark → auto → light', () => {
+    expect(cycleTheme('light')).toBe('dark');
+    expect(cycleTheme('dark')).toBe('auto');
+    expect(cycleTheme('auto')).toBe('light');
+  });
+});
